requirements.txt
argon2-cffi-bindings==21.2.0 
argon2-cffi==23.1.0 
certifi==2024.7.4 
cffi==1.17.0 
charset-normalizer==3.3.2 
colorama==0.4.6 
environs==9.5.0 
grpcio==1.63.0 
idna==3.7 
influxdb-client==1.45.0 
kafka-python==2.0.2 
marshmallow==3.21.3 
milvus-lite==2.4.8 
minio==7.2.7 
numpy==2.0.1 
packaging==24.1 
pandas==2.2.2 
pillow==10.4.0 
protobuf==5.27.3 
pycparser==2.22 
pycryptodome==3.20.0 
pymilvus==2.4.5 
python-dateutil==2.9.0.post0 
python-dotenv==1.0.1 
pytz==2024.1 
reactivex==4.0.4 
requests==2.32.3 
setuptools==72.1.0 
six==1.16.0 
tqdm==4.66.5 
typing-extensions==4.12.2 
tzdata==2024.1 
ujson==5.10.0 
urllib3==2.2.2 
shapely==2.0.5
ollama==0.3.1
slack-sdk==3.31.0
pydub==0.25.1
slack-bolt==1.20.0
flask==3.0.3